requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pyjwt==2.8.0
pydantic==2.9.0
